Completed different echo styles.
Now supporting echo styles of {none,prompt,input,both} to allow for
different styles of programs to be tested seamlessly.  Debugging
required recognizing that the  'read' builtin for bash usually reads
up to a newline but when reading prompts, this is usually not the
case.

This raises the issue space separation for the prompt is currently
baked into testy which is a bit dangerous. For present purposes, I'll
hopefully get away with it.