Reading in extra results from process; eliminating diff spaces.
Extra output is still a bit flaky because there are so many
concurrency issues with knowing when a process is actually ready for
output.  Some of this can be fixed by architecting the tests well.

Added a sed filter on the diff output to reduce the column widths if
possible.  this is a bit risky if the output actually has large spaces
in it but hopefully will not be a major issue.