Partial valgrind support added.
Still looking to do a rerun option but this is starting to shape
up. Valgrind output to a file, checking return code from valgrind to
determine errors detected.

Want to switch to ${parameter:-word} syntax for defaults soon as this
would allow more dynamic setting of some variables and avoid the need
for globals statements.