Fixed end of file no newline bug.
By default, the 'read' bash builtin will report no data at the end of
file but this can be coerced to handle it with the following syntax:

while read -r || [ "$REPLY" != "" ]; do
  ...
done < file

Some students who forgot to put in a newline had this problem.

This problem arose when transforming data based on the
ECHOING="input" where testy is adding back in prompts. It is patched
for now but this modification potentially adds an extra newline which
may be undesirable in some testing situations. This can be avoided by
using ECHOING="both" instead which I should change for future releases
of projects where programs are run in batch mode.