Cleaning up /tmp fifos and files.
Got yelled at by the operator due to not cleaning up pipes and files
in /tmp on CSELabs machines. Have fixed this.

May want to move away from pipes entirely and just use an input file
instead. Essentially doing this now anyway as the testing is not
interactive. Will look into this for a later revision.