First multi tests working.
Have basic infrastructure set up, making steady progress on
formatting, no major stressors yet.

Will want to rethink the single tests as well based on this
understanding and eventually merge the single and multi test running
functions (maybe) to reduce the amount of code used.

Multi tests add failure messages as they proceed and the test
resolution checks the 'state' variable then prints all failure
messages. This is a more elegant approach than the single test run and
I should adopt it there along with some of the functions that I have
used in the multi testing.

Also may wan to switch to using the 'timeout' utility in multi tests
as it does allow for delivery of signals to the proper process.

Probably still some bugs in the multi section where environment
variables are not being honored well and will need documentation as
well.