Added mkfifo checking to prevent errors on WSL
Now testing if FIFOs can be created in local directory or if /tmp must
be used instead.  Aimed to support WSL users where mkfifo fails on
Windows-based file systems but /tmp will work. Tested on a vfat
partition but not on an actual WSL installation.

Adjusted both single and multi support.