Miscellaneous updates bug fixes.
Adjusted sed to allow blank specifications of test output to be
accepted; was a problem with sed and is explained in comments on the
test session processing portion.  Applicable only to single testy
runs, not TESTY_MULTI.

Added a TESTY_DEBUG env var in place of DEBUG as the latter is
sometimes used in other programs.

Broke out VALGRIND_PROG options into separate lines to allow filtering
with external commands in needed.  Was required as old options of
Valgrind don't suport the --keep-debuginfo=yes option

During use of post_filter, use 'eval' as this allows more complex
one-liners like post_filter='grep -v "skip this line"' to be used.