Multi-program testing looks okay
Have verified that TESTY_MULTI session with examples/banter_tests.org,
things work as expected. This isn't rigorous as there is no valgrind
output and generally these are easy to handle however this is at least
re-assuring. Will check that results look good by examining blather
next.

Re-formatted multi output a bit in order to match the conventions in
the single test output. Will want to refactor some of the output
functions for multi testing as well to bring them in line with
conventions used elsewhere.