Parsing of org files done, partial main implemenation
Shifted to a dict of tests keyed on numeric index to allow ~tests[1]~
to be test number one. Am always going to be executing according to a
list of "do these tests" so this makes fine sense, may actually be
expandable as can duplicate the names of tests as keys to allow lookup
on names too (though that would double the size of tests and that'd be
bad). Parsing seems to work fine with the limited testing I've done.

Next steps are to start in on test running code.