Able to run prob1/prob2 tests on F24 216 project 1
Minor parsing error fixes, separating out valgrind output, no need to
do a file in this case, nice that valgrind output is in the original
merged stdout/stderr as it is possible to then output it / access it
to see the inline valgrind messages (possibly, haven't tried).

Various short methods added for pre/post segment cleanup with some
needing to be complete.

Next steps are to work on the diff portion and after that determining
of segments pass, then if tests pass.