Work on diff / alignment problem
Needlemann-Wunsch algorithm in align.py, experiments with side-by-side
diff in testyp.  Honing in on this with a few caveats.

Trouble with dealing with extra blank lines. Don't want this to be a
penalty for students but not sure how to handle the diffing to
eliminate them. Tricky to map back to the original ranges of line
numbers AND preserve the diffing.

The advantage of NW is that similar lines do match better in output
while standard diff just sees them as different.