Added test directory support
Supported via #+TESTY: test_directory="some/dir"

Directory is removed then created for the test

Not able to preserve the directory in use across multiple segments so
will want to correct that via test setup that is not per-segment. 216
S24 P5 Problem 5 Test 5 is a good example of this: requires to
separate setups of the test directory even though both segments could
use the same one.

Minor bug fixes around.