Bug fix on parallel execution
Need to rebuild tests_torun lists as it has dangling refs in it which
are not overwritten by assigning IPC test to main tests[] list. There
was an advantage to tracking just the number of the tests to run in
that this couldn't happen but that makes things awkward otherwise.
modified Suite.run_parallel() to fix this