Cleaner implementation of regeneration via saved preambles
Switched from description / description_full fields to simply saving
the preamble for suite / tests / segments. This is very clean by
comparison to the previous approach as it only requires replacing the
session rather than figuring out non-default values for variables and
such. Will want to keep this in mind for alternative input formats
like JSON: save the original format and simply replace the session
with the actual output.