showfail option added, bug fix for sbs_diff_key
Option to show only failures added to the command line to support
reporting on gradescope.

Student Ghislain reported a bug in python 3.14 which is triggered by
sbs_diff_key not being an initialized field in the Segment class
constructor; adding initialize apparently fixes this for him so I've
done so in the mainline.