Fixed buffering error in process communication
Was invoking Popen() with args that lead to buffering by Python under
the hood; TextIOWrapper objects were used for stdin and that leads to
misalignment with the poll() calls I was doing later. This manifested
as truncated output, particularly some Valgrind output looked
wrong. It wasn't really a Valgrind problem, just that Valgrind's
output appears at the end and the truncation leads to failures.

Have re-worked the Popen() function so that only raw file descriptor
I/O is used with no buffering so that poll() has complete
ownership. This likely leads to more read() calls but more reliable
behavior in the end. It seems to fix the issues observed by some
students.