Minor updates to try to satisfy linter; mainly whitespace
Last attempt to make pylint happy. It is too opinionated about style
issues and after adding the 12th or 13th suppression option to the
Makefile linter options, I'm giving up. The signal to noise ratio is
just too poor for it to be useful. Explored a few other options like
pyflake but it is little better. Not likely to use the Makefile for
linting henceforce.