Progress on isolating problem
Got boondoggled into thinking there was some issue with make but the
true culprit is some kind of incompatibility between os.fork() and the
multiprocessing module. Setting PARALLEL=False and running does fine
but with any number of parallel processes, tests start failing. These
are tests involving bash that have input to the shell and output.