Beginning of bug resolution: dup2() to fd 0
https://stackoverflow.com/questions/30134297/python-multiprocessing-stdin-input
Suggestion in this ancient post that multiprocessing does funky stuff
with stdin; allusions to this in the lib docs.  They screw with stdin
so that it is no longer fd 0 and sys.stdin.fileno() returns something
else.  Doing a dup2(some_fd, 0) will restore stdin for programs that
might read input like bash to get their input. Black magic
bullshit. Have documented this with a minimal example file that shows
hanging behavior and the fix. Implemented the fix in testyp and will
refine code before merging to master branch.