Trailing comments / final END_SRC bug fixes, QUOTE files added
Fixed several bugs that are described in notes with parsing org
files. Tailing comments after a segment are now treated as a COMMENT
segment, one that exists but does not run a program. This new parsing
approach also seems to have fixed the trailing END_SRC problem that
has occasionally required adding a newline to the end of files.

Also added support for #+BEGIN_QUOTE filename sections to dump the
contents into a file.  I'd prefer enabling HERE documents in shell
scripts but this would greatly complicate parsing the session.

Both of the above cases are treated by
- Change the segment.program to be COMMENT or QUOTE
- in segment.run() handle these situations specially
- adjust output in a few other spots to make output sane

The re-using the segment.program field for this purpose feels a little
iffy but is good enough for the moment.