Significant work on docs. Points system hack.
Several examples added, README.org updated to reflect python
conventions. Need to finalize some of these but getting close to ready
for merge / publish.

Wanted to ensure that the points system actually work and allows
individual points to be added set for tests so added a hack to do so
by transferring captured points in segments back up to the
test. Will likely want to smooth this in the future.